% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogetpourpoints.R
\name{autogetpourpoints}
\alias{autogetpourpoints}
\title{autogetpourpoints}
\usage{
autogetpourpoints(reservoir, fac)
}
\arguments{
\item{reservoir}{An sf polygon, with an unstandardised raw reservoir}

\item{fac}{A rast, showing accumulated water flow along river}
}
\value{
An sf multipoint where rivers flow into and out of the reservoir
}
\description{
autogetpourpoints
}
