% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damaoi-package.R
\docType{package}
\name{damaoi-package}
\alias{damaoi}
\alias{damaoi-package}
\title{damaoi: Create an 'Area of Interest' around a Constructed Dam for Comparative Impact Evaluations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package creates an 'area of interest' around a constructed dam. Dams have environmental and social impacts, both positive and negative. Impact evaluations today have no consistent way of specifying exactly where has been impacted by dams, so impacts are measured inconsistently. This package facilitates methods to adjust reservoir polygons to match satellite-observed surface water areas, plot upstream and downstream rivers using elevation data and accumulated river flow, and draw buffers clipped by river basins. This helps to determine the areas which could be impacted by dam construction facilitating comparative analysis.
}
\author{
\strong{Maintainer}: Chris Littleboy \email{chris.littleboy@stir.ac.uk} (\href{https://orcid.org/0000-0002-3293-7987}{ORCID})

Other contributors:
\itemize{
  \item Isabel Jones \email{i.l.jones@stir.ac.uk} [contributor, funder]
}

}
\keyword{internal}
