% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getshinyparams.R
\name{getshinyparams}
\alias{getshinyparams}
\title{getshinyparams}
\usage{
getshinyparams(res)
}
\arguments{
\item{res}{An sf polygon, with an unstandardised raw reservoir}
}
\value{
An parameters to start the shiny app for determining pour points manually for a given reservoir
}
\description{
getshinyparams
}
