% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointstolines.R
\name{pointstolines}
\alias{pointstolines}
\title{pointstolines}
\usage{
pointstolines(riverpoints, espg)
}
\arguments{
\item{riverpoints}{list of dataframes returned by the riverpoints function.}

\item{espg}{espg code of UTM zone}
}
\value{
An list of upstream lines (multilinestring) and downstream line (linestring)
}
\description{
pointstolines
}
