% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{preprocessing}
\usage{
preprocessing(
  reservoir,
  dem = NULL,
  fac = NULL,
  water_bodies = NULL,
  basins = NULL,
  pourpoints = NULL,
  river_distance
)
}
\arguments{
\item{reservoir}{An sf polygon, with an unstandardised raw reservoir}

\item{dem}{A rast, showing elevation}

\item{fac}{A rast, showing accumulated water flow along river}

\item{water_bodies}{A rast, where 1 indicates water, NA otherwise}

\item{basins}{An sf multipolygon, with the basins in the area around the dam}

\item{pourpoints}{An sf multipoint, showing the points where rivers flow in and out of reservoirs}

\item{river_distance}{A number, indicating the number of meters downstream and upstream for the area of interest. Defaults to 100000 (100km)}
}
\value{
A list with utm transformed input data
}
\description{
preprocessing
}
